core/requirements.txt
Django==3.1.2
django_rest_swagger==2.2.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
gunicorn==20.1.0
